fcasd

+++ app/Http/Controllers/CartController.php
@@ -35,28 +35,43 @@ class CartController extends Controller
         ]);
     }
     public function store(Request $request){
-        dd($request->all());
         if (Auth::check()){
-            $addCart = Cart::create([
-                'product_id'=>$request->input('productId'),
-                'quantity'=>$request->input('quantity'),
-                'user_id'=>Auth::user()->id,
-            ]);
-            $addCart = Checkout::create([
-                'product_id'=>$request->input('productId'),
-                'quantity'=>$request->input('quantity'),
-                'user_id'=>Auth::user()->id,
-            ]);
-            return redirect(url('cart'))->with('success', 'Item Added to Cart Successfully');
+            $getProductId =Cart::where('product_id',$request->productId)->where('user_id',Auth::user()->id)->first();
+            if (isset($getProductId)){
+            $productId = $getProductId->product_id;
+                $getQuantity = Cart::where('user_id',Auth::user()->id)->where('product_id',$request->productId)->first();
+                $quantity = $getQuantity->quantity;
+                $updateQuantity = $quantity + $request->quantity;
+            $updateCart = Cart::where('user_id',Auth::user()->id)->where('product_id',$request->productId)->update(['quantity'=>($updateQuantity)]);
+            $updateCart = Checkout::where('user_id',Auth::user()->id)->where('product_id',$request->productId)->update(['quantity'=>($updateQuantity)]);
+
+                return redirect(url('cart'))->with('success', 'Item Added to Cart Successfully');
+
+
+            }
+            else {
+                $addCart = Cart::create([
+                    'product_id' => $request->input('productId'),
+                    'quantity' => $request->input('quantity'),
+                    'user_id' => Auth::user()->id,
+                ]);
+                $addCart = Checkout::create([
+                    'product_id' => $request->input('productId'),
+                    'quantity' => $request->input('quantity'),
+                    'user_id' => Auth::user()->id,
+                ]);
+                return redirect(url('cart'))->with('success', 'Item Added to Cart Successfully');
+            }
         }
         else{
             return redirect(url('login'));
         }
 
+
     }
     public function delete($id){
-        $deleteCart = Cart::where('user_id',$id)->delete();
-        $deleteCheckout = Checkout::where('user_id',$id)->delete();
+        $deleteCart = Cart::where('product_id',$id)->where('user_id',Auth::user()->id)->delete();
+        $deleteCheckout = Checkout::where('product_id',$id)->where('user_id',Auth::user()->id)->delete();
 
 
         return redirect()->back()->with('success','Item Removed From Cart');
